refactor: move createCars from utils to components
